Add option for smoother labels rendering in 3D (#7100)
# References and relevant issues
There have been a few relevant discussions on Zulip recently about
rendering labels in 3D.

https://napari.zulipchat.com/#narrow/stream/212875-general/topic/smoother.20labels.20volume.20rendering

https://napari.zulipchat.com/#narrow/stream/212875-general/topic/smooth.20image.20in.203d

https://napari.zulipchat.com/#narrow/stream/212875-general/topic/better.20labels.20rendering

# Description
This PR adds a new option designed to provide smoother volume rendering
of labels. It works by using a larger and more isotropic (3D Sobel)
kernel when the isosurface shader is calculating the local gradient of
the labels volume, which it uses to set a normal vector.

Here is a comparison from the `examples/labels3d.py` example:

![smoother-labels-side-by-side](https://github.com/user-attachments/assets/adf7f3cd-2d56-4252-8840-5ad44cb1892a)

And here using a mask from the Vesuvius challenge data:

![smoother-labels-toggle](https://github.com/user-attachments/assets/26c2e931-d462-453d-bc81-8337df018f04)

@JoOkuma proposed a nicer looking implementation that uses a signed
distance function calculated on the CPU before rendering. The
conversation there seemed to indicate that would take longer to
implement properly in napari, so I proposed this as a short-term
improvement that is still a decent improvement to my eyes.

---------

Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>