Disconnect all dims events when closing viewer (#7140)
# References and relevant issues
Closes #6685 

# Description
This disconnects all the `Dims` events from `Viewer` methods when
calling `Viewer.close`. This prevents the `Viewer` from responding to
changes to `Dims` as layers are removed one-by-one.

It also fixes another related TODO and some typos.

After some consideration (documented by the comments in #6685) I decided
that this approach is justified because `Viewer.close` has lots of
side-effects that make `Viewer` and things it owns mostly unusable.
Compare to other simpler apporaches, this also avoids unnecessary work
when closing the viewer.

I wonder if we should disconnect even more events when calling
`Viewer.close`, but I'll leave that for future work.