FIX `QtViewer._open_files_dialog` handing of `stack` (#7172)
# References and relevant issues

closes #7165

# Description

Correctly handles `stack` parameter in `_open_files_dialog`. This was
due to a bad refactor in https://github.com/napari/napari/pull/4865

I noticed that we don't test any of the `QtViewer` `open_files...` or
`open_folder...` methods. (Some are tested with the file menu but none
of these seem to be tested in isolation?)
I've added a new test file for these but only added a test for this
instance. Maybe additional tests can be added in a follow up PR, in
order to get this fix in quicker?