Update shortcuts.py to have `Enter` be primary Shapes completion (#7063)
# References and relevant issues
Part of: https://github.com/napari/napari/issues/6489

# Description
For the Labels polygon tool, Enter completes the Shape and Escape
aborts.
The Shapes layer doesn't have an abort, but this PR adds Enter as the
primary shape finish/complete shortcut to match Labels. The shortcuts
are edible by the user, but this way out of the box there will be more
consistency.