Speed up highlighting in the Shapes layer (#7223)
# References and relevant issues

This PR fixes highlight performance problems in the Shapes layer. #7162
tried to fix them by reducing the frequency of highlight events, but
that had to be reverted (#7187, #7201). Instead, this dramatically
speeds up highlighting when the event is fired.

# Description

For performance reasons, the Shapes layer creates a *single* mesh
containing all shapes and shape edges, and this is what is sent to VisPy
to draw.

When highlighting, prior to this PR, we needed to find the Shape edges
that are being highlighted within that giant mesh (linear in the number
of vertices, and iterated in a Python for-loop).

With this PR, instead, we look at the precomputed mesh stored in each
highlighted shape and stack them together. Therefore, the complexity is
linear in the number of vertices in the highlighted shapes, and the
constant factors should also be reduced.

Before this PR:
<img width="343" alt="Screenshot 2024-08-29 at 11 50 34"
src="https://github.com/user-attachments/assets/bf2fb01d-494a-42c4-8f06-53dc14dbac0b">

After This PR

<img width="273" alt="Screenshot 2024-08-29 at 11 50 43"
src="https://github.com/user-attachments/assets/52720dad-19bc-42c4-949d-e62bd5544e34">