Fix convex polygon detection (#7228)
In reviewing #6654, I discovered that the `_is_convex` helper function,
implemented in #7214, wasn't working correctly, because the coordinate
arrays passed to the `orientation` function were transposed relative to
what that function expects. (It was a coincidence that earlier examples
worked in either order.)

I've transposed the arrays being passed to the function and updated the
tests to match.