Sort plugin manifests in alphabetical order for registration (#7266)
# References and relevant issues
Closes #5500

# Description
This PR ensures that when the viewer is launched, plugin manifests (and
therefore their actions) are registered in alphabetical order of
manifest's display name (if available) or plugin name.

Note:

- if a plugin is installed/enabled while the viewer is launched, this
plugin will go to the bottom of the list, not in its correct spot. It
will be correctly sorted on viewer relaunch. See [zulip
thread](https://napari.zulipchat.com/#narrow/stream/212875-general/topic/sorting.20plugins.20menu)
for more discussion on this.
- Plugins that offer single commands will still be sorted in plugin name
order, not action name order, and that may still be confusing. This is a
non-trivial fix. Will continue to look at that.

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>