Enforce minimum side length when guessing if image is RGB (#7273)
# References and relevant issues

Closes #7261

# Description

For the vast majority of user RGB images, the side length would be
larger than 30 pixels. (For reference, our layer thumbnails are 32x32.)
Therefore, we can reduce or avoid issues like #7261 by enforcing a
minimum side length before we use the final dim as an RGB dim.

I've implemented this as a parameter on the guess_rgb function, which
allows us to turn this into a preference in the future.