Enh: Add zoom options and keybindings to View menu (#7200)
# References and relevant issues

Closes https://github.com/napari/napari/issues/4855

# Description

This PR adds menu items and "standard keybindings" for Zoom in, Zoom
out, and Reset Zoom to the View menu.
The keybindings are Qt defaults
(https://doc.qt.io/qtcreator/creator-keyboard-shortcuts.html#image-viewer-shortcuts),
as well as defaults on macOS and Windows (see e.g. Adobe Photoshop or
Chrome)
Note: I've made a PR to app-model to add the Reset Zoom standard
keybinding (https://github.com/pyapp-kit/app-model/pull/220) so if that
is merged I can edit this PR to use that here too!

I added to a test for the new behavior.

Note, this approach does not present these as editable in the napari
Preferences. I
These are bound like an OS default would be, e.g. copy, paste, etc. They
are editable on the OS level at least on macOS.

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>