Eliminate nearly all Qt widget leaks by using qtbot (#7251)
# Description

This PR is a next PR to clean up bugs in tests that are causing
segfaults for some test run. Previous one was #5373

In this PR I have focused on finding widgets that are not properly
teardown.

Next to the improving of tests:

1) change part of the test to use `QtViewer` instead of full `Viewer` 
2) add missed widget to qtbot 
3) use `before_close_func` instead of custom teardown code in fixture


This PR contains a code for detecting of leaked widgets. This code
contains two parts.

The own subclass of QtBot that ensures that object, with is hadled using
`addWidget` name is set to `'handled_widget'` affter the test end.
And `_find_dangling_widgets` fixture that is executed after the test end
(and after qtbot teardown) and looks for all top-level widgets that do
not have parents and check if this widgets are either:

1. `_QtMainWindow` - we handle it using `make_napari_viewer` fixture,
without using qtbot.
2. some widget form qtconsole (I do not have time to investigate if we
could make a PR to qtconsole to avoid it).
3. is not `CanvasBackendDesktop` that is leaking in `test_sys_info.py`
and may require PR to vispy to fix this. (or some monkeypatching)
4. is not handled by qtbot already by having name `"handled_widget"` 
5. add fixtures for `viewer_model` and `qt_viewer` 

If the widget does not fit to any above category, it is reported as an
error.

There is no guarantee that all problematic widgets are found. So if this
PR will be merged then some random test failures need to be observed, as
some widget not too often leak from tests.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>