Speed up `get_status` for Shapes layer by using bounding boxes (#7144)
# References and relevant issues

# Description

When investigating why napari interface is so lagging with big number of
layers, I have spotted that Shapes triangulation is used to determine if
the mouse is inside some shape. However, a single shape could be
constructed from multiple polygons. Also, edge polygons are used.

This PR adds information about bounding box to Shape object. Then
filtering of Shapes based on belonging cursor position to bounding box
of shape.

Then checking if the cursor is in such a filtered list of shapes is
performed based on triangles (same as previous).

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>