Fix color shuffling for bool labels (#7294)
# References and relevant issues

Does not quite deal with #7277 (but sort of does)

# Description

In investigating #7277 I found that:

- we automatically convert bool volumes to int8
- we do this with a *copy*, not a view, which is bad because then editing the
  Labels layer does not change the input array (which we do for every
  other data type)
- our logic for shuffling colors for signed integers is faulty — we assign
  colors with `arr % n`, where `n` can be larger than the maximum value
  for the dtype, causing an `OverflowError`.

This PR does **not** address the third point, only the first two, by:

- using uint8 instead of int8 in the case of bool labels, since sign is
  meaningless here.
- using `data.view` instead of `data.astype`, which ensures that
  painting into the data also paints into the input array.

I've added a test for that behaviour.

I think it's worth getting this in for 0.5.4, and dealing with the broader
problem of signed ints in subsequent releases, because bool masks are very very
common, while signed int8 and int16 are very *un*common. (And this problem does
not affect larger signed integers.)

As a side note, I was convinced this dated from our scramble to get Labels
overhauled for 0.4.19, but in fact it dates all the way back to #2496! 😂