fix: remove "openblas_set_num_threads not found" warning by add missed break (#7369)
# Description

When start napari on macOS that uses `numpy<2` its emit warning
"openblas_set_num_threads not found" even if it was found, and number of
threads is set properly.

This is because of a missed break, so for loop always do all loops and
ends in `else` clause. This PR fixes it.