Isolate keybinding tests using pytest.mark (#7375)
# References and relevant issues
Refactoring keybindings.

# Description
This PR uses the pytest mark functionality to identify tests that are
related to keybindings. As things
with keybindings are refactored, this would speed up local testing of
keybinding functionality.

To run the keybinding tests, enter:
```sh
pytest -v -m key_bindings
```

cc/ @lucyleeow You may find this useful. If so, I will take this out of
draft.