Add example doc comparing `screenshot` and `export_figure` (#7319)
# Description

The current examples `to_screenshot.py` and `export_figure.py` are not
immediately clear on the differences and limitations of each. I have
combined each document into one new example, with added comparisons to
clearly show how margins are handled. Furthermore, this example
clarifies that overlays like `scale_bar` are also exported with the
figure and dependent on the margins, which is not immediately obvious
since a canvas view can have the `scale_bar` outside the margins of the
data. This example clearly shows that the scale bar is moved to fit
within the extent of the data.

I believe that combining the examples in this way will improve user
understanding of these two features, and make searching for a screenshot
without margins feature easier. I would also be interested in removing
both `to_screenshot.py` and `export_figure.py` since this example
contains clarity on functionality of both, such as the scale bar, that
is not present in the prior files (though this could be added). Also
note that the original docstrings for both files are incorrect and
should otherwise be updated.

Finally, this example shows how the scale bar length is dynamically
adjusted with calls to `export_figure` which may be confusing for users.
I'm not sure if adding `viewer.scale_bar.length = 150` would make this
functionality more or less clear to the user, so I've added it as a
comment for now.


![image](https://github.com/user-attachments/assets/73f6f881-986a-4811-b903-54739a23598f)

---------

Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>