mark slow local tests (#7383)
# References and relevant issues

Partially addresses #7382

# Description
This PR marks the 10 slowest tests on my local machine. This allows me
to include or exclude when I run tests locally.

This could be an initial proof of concept step that could be used to
identify some slow
tests on CI runs and treat them differently than the rest of the test
suite. Perhaps exclude on some test matrix runs or parallelize the slow
tests.

---------

Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>