Turn off 'flash' in `screenshot`/`export_figure` examples (#7427)
# References and relevant issues

Addresses / Closes? #7425 

# Description

Examples using `screenshot` and `figure_export` were found to have a
grayscale canvas in the docs. The grayscale appears to be from the
`flash=True` argument, so this PR sets `flash=False` to fix the display
of the viewer for the Gallery. The canvas will now properly be color.

This does not fix the root cause of the grayscale canvas as a result of
`flash=True`.

Tested with a local docs build (finally!)

Thank you @psobolewskiPhD for assistance.