Elide layer name in the middle instead of the end (#7461)
# References and relevant issues

closes #7460

# Description

Layer names can often be long, and the distinguishing characteristic is
as often at the end of the layer name as at the start. When a layer name
is too long to display, we should therefore put the ellipsis in the
middle of the layer name, rather than at the end.