stop/start notification timer on window focus change (#7392)
# References and relevant issues
https://github.com/napari/napari/pull/7220#issuecomment-2477654620

# Description

This PR stops the notification timer if windows lost focus, and starts
the timer for the top one most after gaining focus again.

If a new notification shows, then the previous one timer is stopped and
started again if it is closed.

The notification timer is restarted after the mouse leaves the dialog
area.

I do not handle the case of hundreds of notifications. Maybe someone
will have an idea how to do this properly.