Extend reading with plugins to allow Layer objects (#7443)
# References and relevant issues
Addresses issue reported in this [zulip
comment](https://napari.zulipchat.com/#narrow/channel/309872-plugins/topic/Reader.20plugin.20with.20connected.20event/near/485161161)

# Description
Extends `add_layers_with_plugins` to handle instantiated `Layer` objects
returned from `reader` plugins, not just `LayerDataTuple`s.

This allows plugin developers to instantiate layers and connect
callbacks to them (if desired) before returning from the reader.

Note: this plugin sets the source on the returned `Layer` if and only if
the plugin developer has left both `path` and `reader_plugin` of the
Layer's source blank. We may want to **always** override the source of
the layer, if we judge this is likely to lead to better UI. In order to
set the source of an already created layer, I had to add a `source`
setter, which is currently public. We may want to guard it, so that
users/plugin developers can't override a `Layer'`s source and it's
entirely within our control?