Specify dtype when using zarr.Group.create_array (#7497)
Closes #7496

It looks like create_array used to have a default dtype in zarr-python 2.x but
now the dtype is mandatory. I'll make a corresponding issue in the zarr-python
repo but in the meantime fix the issue locally by specifying the dtype — I kind
of agree that it should be a required argument anyway.