Flip z axis on 3D camera to default to right-handed frame (#7488)
Closes #4633

This is one of those embarrassing moments where a fix was literally one
line
but took ~~2.5y~~ 4y to fix. 😅

After investigating the lat/lon display issue for some xarray data
([community meeting
notes](https://hackmd.io/9vo_DiJpSfuUExeGqVHbDQ#2024-11-27)),
@melonora helped me realise that we can flip the display axis directions
trivially in the VispyCamera model (and indeed already do so in 2D to
get
axis 0 to point down).

This PR adds a flip to z in the 3D camera so that the 3D frame is
right-handed
by default.

I intend to make a follow-up PR adding an API to the Camera model so
that users
can control this. However, I do think that the default should be a
right-handed
frame, as most users expect this. My impression is that 99.9% of people
don't
care about the handedness of their data, and the remaining 0.1% are
split in
similar proportions between those that expect a right handed frame and
those
that expect a left-handed one. To wit, both the molecular models
demonstrating
the issue in #4633 and medical image data I opened with nibabel were
mirrored,
as well as the user's data from [this image.sc
post](https://forum.image.sc/t/3d-view-coordinate-system-is-left-handed/66995).