[maint] Update pyproject.toml to move plugin manager to optional and let triangle work on arm64 (#7511)
Currently the only way to install napari with the plugin-manager is via
`all` which is `pyqt5`.
This PR moves the plugin-manager to `optional`, so that you can at least
use `pip install "[pyqt6, optional]"` to get a functional napari and the
plugin manager.
For those using `[all]` nothing will change, because `[all]` includes
`optional`.

Alternately, it could go in the base requirements along side
`napari-svg`...

While doing this, I noticed the exclusion flag of arm64 for triangle.
A recent release has wheels now for just about everything:
https://pypi.org/project/triangle/#files
so this flag can be removed.