Add poly line drawing (#7099)
# References and relevant issues
closes #7080 

# Description
Currently napari has the capability to add a path to the shapes layer,
but it does not allow for adding a more freehand drawing of a line in
the shapes layer. One could do this using the labels layer, but the
downside of that is that the size of the label layer being created is at
least equal to the current extent of layer data in the viewer. This PR
adds the capability of adding a polyline to the shapes layer, very much
like freehand drawing of a line. This functionality works both by using
the mouse or a pencil on a tablet. A shortcut using `Shift+L` is
provided when the `shapes` layer is active.

This work is the result of a napari code cafe.

---------

Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>