Iterative ROI Screenshots (#7209)
Motivation

When drawing rois using shapes layer and taking screenshots, the user is
required to manually center a particular roi and adjust the zoom level
to allow for this. After having drawn multiple ROIs, perhaps even across
multiple samples a user might quickly want to generate all the
screenshots. This PR allows for that.

:rocket: Feature
Similar to the way the export_figure function works, this function
iteratively centers the camera to the center of each ROI and adjusts the
size of the canvas to the size of each ROI and takes a screenshot, and
optionally saved the ROI as an image. Lastly, the previous camera
position, zoom level and canvas size is restored.

The function can take in a `list` of paths, the length of which must be
equal to the rois provided. Providing a `str` or `Path` is also allowed
in which case a directory will be created if it does not exist already.
Rois are then stored as `roi_{n}.png`. The rois should be provided as a
list of arrays with shape `(4, 2)` representing a rectangular roi. If
one roi does not match this shape an error is thrown that the shape is
invalid. This error also includes the message that 3D is not supported.
Using `export_rois` is also not supported in 3D view, an error will be
thrown if a user attempts this.


<img width="1189" alt="Screenshot 2024-08-19 at 11 51 40 AM"
src="https://github.com/user-attachments/assets/3f13dff2-5bbc-49bb-8fde-977f4da434a5">

<img width="636" alt="Screenshot 2024-08-19 at 11 43 14 AM"
src="https://github.com/user-attachments/assets/0d95e0d7-e65f-4b0f-919b-59e4979fceba">

---------

Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>
Co-authored-by: Wouter-Michiel Vierdag <w-mv@hotmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Wouter-Michiel Vierdag <michiel.vierdag@embl.de>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: Draga Doncila Pop <17995243+DragaDoncila@users.noreply.github.com>
Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>