Use data.dtype for creating slicer (#7540)
# References and relevant issues
closes #7535

# Description

When async mode is on, the async image slicer starts by creating an
empty slice. Before this PR, the slice had type np.uint8. For images,
this was fine, because when the real slice arrived the dtype got updated
correctly. For labels, though, the incorrect dtype caused an incorrect
caching of the data dtype -> texture dtype colormap assigned to the
layer, which resulted in incorrect colors being displayed.

This PR ensures that the empty slice dtype matches the image dtype.