Add right-click indicator to 3D, Roll, Grid, and Square push buttons (#7556)
# References and relevant issues

Addresses adding special indicator to GUI buttons that have right click
menu's in #7502. Prompted by discussion in #7485.

# Description


![image](https://github.com/user-attachments/assets/e1272752-8994-4227-8cd6-a798f98a98a0)


Changes icon for Roll, Grid, and Square push buttons to include a
caret-like symbol in the lower right corner to signal visually the
ability to right click this button for extra menus. Consistent with
behavior in other programs.

1. Roll icon (re)-centered to accommodate for indicator
2. Grid icon made smaller to allow room for indicator
3. Square icon unchanged except for indicator addition. 

Tested to work in both light and dark theme. (Note to future icon
editors, the `undefined` fill and stroke is required for proper theming,
but this behavior is inconsistent, at least with Inkscape v1.4).