Add right click indicator to playback icons (#7590)
# References and relevant issues

#7502 started the work of adding special menu indicators to buttons.
After #7556 I noticed that 'square.svg' was shared for both the
'non-grid' view and the stop playback button. So this both fixes that
crossover from main, and adds the feature intentionally.

# Description

1. Adds indicator to playback buttons for dim sliders, but giving new
file names (and keeping old files for other usage).
2. Removes indicator from 'square.svg'. I believe this is no longer
used, but I left it in
3. Updates styling. Notice how on main the stop and play buttons were
different sizes.

Bug?? The 'playback-reverse.svg' or 'left_arrow.svg' is never used. It
will always use the forward button, even if the mode is reversed.
Doesn't effect my understanding, but thought I'd point it out.

This PR: 

![python_X0TZTfe6p8](https://github.com/user-attachments/assets/3de98cb6-933a-4fd8-ab81-61774fcc8c5d)

Main:

![python_gNJkcDr0J7](https://github.com/user-attachments/assets/14fd09c2-8178-41cc-a6ff-285421a0f01a)