Add Grid Mode Spacing to change distance between layers  (#7597)
# References and relevant issues
#7587 includes the desire for the ability to add spacing between layers
in Grid mode. At the Feb 5/6 meeting, we discussed adding this ability,
and frequent grid mode proponents @DragaDoncila and @psobolewskiPhD may
be especially interested in this PR. This spacing especially helps quick
export of figures, and no longer having to precisely crop panels 😄

# Description


![python_kcX7tXeNAw](https://github.com/user-attachments/assets/3a06c12a-b7db-4d31-9139-e05f11b3e201)

1. Adds proportional spacing between layers in grid mode. Distance is
calculated as the average of the horizontal and vertical layer extent
during the creation of the subplot. float between [-1, 1]. By using the
float value the user is less likely to confuse a larger % value (like
80) as pixels vs percent
2. Accounts for spacing in `viewer.reset_view()` to properly center
camera
3. Adds spacing value to Grid GUI widget. Inherits preference value
initially. Step size is 0.05 (aka 5%).
4. Adds spacing value, by default 0, to preferences. Personally, I'm
going to set this to 0.1, because I really enjoy the breathing room
between layers. I *propose* to also make this the default value, but
that would greatly change things for people!
5. Adds custom type `confloat` from pydantic that accepts step size.
This is then used to allow a 0.05 step size value for the preferences
widget. Type usable for future float widgets!
6. Adds tests to ensure GridCanvas properly set up and that the spacing
of the layers is properly calculated (in test_viewer_model)

Notes:
- [ ] I don't see a place where `viewer.reset_view()` is tested, which
would be good to ensure camera centering is correct. I have validated by
hand that it works AND it works properly with `viewer.export_figure()`,
so does not break anything as far as I can tell
- [x] Currently, allows setting a negative spacing value, which would
result in overlap. I think this could be useful in some circumstances,
but it may also make sense to set the spacing values as [0,1]. One use
case I can think of is display the same image with a variety of
convolutions, that is like 20 different layers, and have them be much
closer in space to compare the overlaps, especially if grid stride one
and blending is translucent
- [x] Fixed with custom `confloat` type~The preference setting jumps by
1.0 increments, I would prefer also to have the 0.05 step size as used
by the GUI widget, but can't figure out how to do this (only validating,
not stepping)~
- [x] Added, for the first time, `.. versionadded::` to a docstring in
the napari repo lol. It does seem to apply, but no one has previously
used it. We will find out if docs build it nicely on release. ~I don't
know if this applies in the Final Checklist - I don't see this in code
docstrings, so is it just changed in napari/docs? I thought the api docs
were auto created:~
- ~If an API has been modified, I have added a `.. versionadded::` or
`.. versionchanged::`
  directive to the appropriate docstring (For more information see
[the Sphinx
documentation](https://www.sphinx-doc.org/en/master/usage/restructuredtext/directives.html#describing-changes-between-versions)).~

---------

Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>