Change ndisplay button to toggle-like to increase discoverability (#7608)
# References and relevant issues

Closes #7486. It is hard to know that the 'square' icon for 2D leads to
3D mode if it's clicked on.

# Final PR Description

Adds a wider button which concurrently visualizes both 2D and 3D
ndisplay modes. Uses a visual prioritization and linear gradient using
the theme highlight to show the current ndisplay mode, and a small
indicator with a smaller icon to show the alternative. Names the new,
wider vector files to '-toggle' and preserves the original icons in case
downstream packages rely on their appearance


![image](https://github.com/user-attachments/assets/6579f84a-0150-481a-a311-e27e7907d0c2)

![python_HVcwDVzCSj](https://github.com/user-attachments/assets/b7e152d6-8eb5-4643-baa5-c3abf55a1ef4)

# Description

Original PR: 

![python_knPBfEiHOQ](https://github.com/user-attachments/assets/ebb00c52-2bfa-4653-ac22-a8568ded6e94)

1. Change styling of ndisplay button to be 75% wider
2. Change 2D and 3D buttons to reveal what's underneath
3. Name these buttons 2D-toggle and 3D-toggle to preserve anything that
utilizes the original 2D and 3D buttons (like in other packages)