Enable creation of custom linear colormaps in layer controls (#7600)
- replace QLabel with QPushButton in the image controls

# Description

This PR replaces the 'colormap' label (which displays the current
colormap) with a push button (which also displays the colormap as
before).
When the button is pressed, it will ask the user to select a color using
color picker.
If a color was selected, then it creates a new colormap, otherwise
nothing happens.



https://github.com/user-attachments/assets/81771c13-cba9-4fec-851d-327748fc2517

---------

Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>
Co-authored-by: Wouter-Michiel Vierdag <w-mv@hotmail.com>