Fix layout issue in image/surface controls (#7618)
This fixes a minor layout issue for the image/surface layer controls
(which use. the colormapComboBox.

I noticed that if the width of the layer controls is changed, the
'colormap' combobox doesn't get resized (because there was a stretch
after the control).

I should have noticed in #7600 but clearly missed it.

before

![image](https://github.com/user-attachments/assets/a5644564-274b-4ba6-b522-c2639a22fed4)

after

![image](https://github.com/user-attachments/assets/7ec326c8-3441-4be2-9dbb-3c4e293fef78)

# References and relevant issues

# Description