Add optional dependency sections for gallery and docs (#7487)
Looking for feedback here, especially from @Czaki and @melissawm.

Prompted by [this
comment](https://github.com/napari/napari/pull/7468#issuecomment-2564866413)
from @psobolewskiPhD.

When adding an example to the gallery, any new dependencies need to be
added to
both (currently nonexistent) `docs` and `testing` optional dependency
groups.
That made me think that there was something a bit off with how we were
thinking
about things. I thought about this a bit more and thought it best to
create a
"gallery" group, which both testing and docs depend on.

I've implemented this in this PR, adding [docs] and [gallery] and then
adding
dependencies on [gallery] to both [docs] and [testing]. Some questions:

- Currently, I've added anything that is imported in the gallery to the
gallery
dependencies, even if it is already a core dependency. The rationale
here is
that if we, for example, drop the dependency on pandas, we still want
the
examples using pandas to run. It's nice to have our core dependencies
and our
  *direct* example dependencies uncoupled.
- However, adding minimum versions to the examples makes it quite
onerous to
keep gallery dependencies in sync. I've therefore kept gallery
dependencies
  bare, with no version requirements. Is that a bad idea?
- The docs dependencies are copied over from [the requirements.txt file
in the
docs
repo](https://github.com/napari/docs/blob/45d18dcd4653f56e9fdaf438a5897882c44fc13a/requirements.txt).
  Is that what we want to do or do we want to clean these up/add
minimum/maximum versions? (And maybe that could come in a separate PR?)

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>