Fix add_shapes.py example: remove duplicate vertex that can cause seg faults (#7636)
# References and relevant issues
Closes: https://github.com/napari/napari/issues/7634
(though the underlying root cause still needs to be fixed)
See https://github.com/napari/napari/issues/7634#issuecomment-2676057213
for discussion

Should also fix docs runs, e.g.
https://github.com/napari/docs/actions/runs/13459251801/job/37610304656

# Description
One of the polygons in the example has a duplicated vertex.
This can cause seg faults in triangulation, which is happening in the
docs right now.
So in this PR i remove the duplicate -- it doesn't affect the visual at
all and it's not the point of the example.