Expose additional Camera parameters in GUI with 3D popup widget (#7626)
# References and relevant issues

Task in #7587 to expose more canvas settings in the GUI. This exposes
zoom and angles of the Camera in the popup widget

# Description

A user wanted to reproduce the zoom and angle of the 3D camera between
sessions, but with the desire to not use the console. I'd think this is
not uncommon. This PR adds multiple widgets to the ndisplay==3 popup to
expose these settings, and adds zoom to ndisplay==2. It adds a labeled
slider widget to all camera settings. Adding this in the form of a
widget also increases accessibility for users with difficulty operating
a pointer for spinning. Open to discussion about step size and number of
decimals displayed. Because this widget popup cannot be used at the same
time as the canvas itself, I don't think too many callbacks need hooked
up.

Current behavior is to bound angles with X[-180, 180], Y(-90,90) and
Z[-180,180] because values outside (-90,90) for Y result in
recalculation of viewer.camera.angles. See [this comment
chain](https://github.com/napari/napari/pull/7626#discussion_r1966955489)
for discussion of this implementation.

Current PR:

![image](https://github.com/user-attachments/assets/eb553e41-a3d0-40d9-a303-7896e2a1c7c6)

## Original Description
This PR adds multiple widgets to the ndisplay==3 popup to expose these
settings. It also changes the perspective slider to a spinbox; however,
if this is well received it would be nice to update each of these
widgets to include a slider (for easy spinning) _and_ the spinbox.
Adding this in the form of a widget also increases accessibility for
users with difficulty operating a pointer for spinning. Open to
discussion about step size and number of decimals displayed. Because
this widget popup cannot be used at the same time as the canvas itself,
I don't think too many callbacks need hooked up.


![image](https://github.com/user-attachments/assets/456e549b-05e0-44e8-82ee-75b91818a1b0)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>