Add version info to welcome screen (#7659)
# References and relevant issues

Closes #7658

# Description

1. Adds version info (e.g. `napari 0.6.0`) on welcome screen between
logo and 'Drag image(s)...' label.
2. Sets smaller font size with qss

Open to suggestions on where it should be located (it makes sense above
the drag to me because that ends in a colon. It does not look great not
aligned to center because it gets placed all the way on the edge of the
canvas.

Could put below shortcuts? 

Would look like this with 0.6.0:

![image](https://github.com/user-attachments/assets/93c793a8-a71e-4aad-af6c-1a6cf77a07d9)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>