Add 'Extend with Plugins' and 'Contribute to napari' links to the Help menu. (#7645)
# Description

Adds two new links to the Help Menu. These are opinionated additions for
sure.

1. 'Plugin Usage' currently links to the Plugin index, but perhaps it
would be nice to link to the hub instead? My thought was by going to
plugin usage you would learn, then, how to find the hub, but also look
through pip and conda stuff. Also, this links people to the how to make
plugin docs ;)
2. 'Contribute to napari' links to the Contributing docs. I think this
might be a bit of encouragement for any user to check out and learn
more. Keeps this link even when in dev because might be a useful
reminder to someone if they are working on contributions, regardless.